Split Android review tests
